Filter out irrelevant automake build targets
Makefiles generated by automake include a number of targets intended for
internal use, and other convenience targets which are not productive
build targets. (eg, *-am, *-recursive, maintainer-clean, installchecks,
etc)

The full list of available targets would frustrate a
productivity-inclined user.

Some rarely used, but useful targets are kept, such as:
- install
- uninstall
- config.status, Makefile, maintainer-clean
- check, distcheck
- dist
- html, pdf, dvi

In the future, a tiered selection could be implemented, where the
available targets are classified as primary, secondary and all, if
automake's capabilities need to be fully exposed by the makefile tool
extension. Such a selection may avoid frustration while still making
it possible to select rarely used targets.

This commit implements a filter to remove targets which are not suitable
for inclusion in the list of available targets for selection.