Make buildLog per-configuration
Previously, one buildLog was used for all configurations, which is
wrong, because the build is subject to configuration.

When a (non-Default) configuration is active, its buildLog will be used.
If not set, no buildLog will be used. The Default configuration buildLog
will not be used for a specific configuration.

When the apropriate buildLog (either Default or specific) is not set,
the make --dry-run process will be followed to obtain a fresh
build configuration.