Move dryrun.log to builddir - optionally saved
Saving the dryrun.log to the extension output directory doesn't extract
much value from it, especially as it is overriden per configuration.

This commit makes saving it to the disk optional (default = not saved),
configurable with a new makefile.keepDryRuns option, and moves the save
location to ${makeDirectory}/.vscode/${targetName}.dryrun.log

This location allows changes to the configuration to be easily tracked.
For example, it could be added to git for easy review of configuration
changes.

If the build directory is out of tree, you could create a git
repo specifically for tracking configurations of builds.

If the srcdir and builddir are the same (in-tree building), you could
use the source repo to also track builds; mixing source and diagnostic
files can be messy; a separate repo initialized in the .vscode directory
can help. Otherwise, worktrees or hard links can be useful to track the
dry-run output in a separate repo.